refactor: split check/embed CLI helpers and path modules
Narrow the embed, check-rendering, and path roots into explicit helper modules while preserving existing public patch points and command behavior.
